Move sales page methods and properties to a new trait.
